Reworked Player control method + SpriteAnimatorComponent extension
Added PlayerControllerComponent that pulls the input and then calls Move() or Jump() on the PlayerScriptComponent. I did this because in the future for my enemies they will need to be able to be controlled by an AIControllerComponent or a PlayerController
SpriteAnimatorComponent  is now also responsible for storing all animation clips. I previosuly stored these in the PlayerScirptComponent but that didnt make any sense.